linux-firmware-rpidistro: Guard the recipe under a license flag
We do this as a result of an upstream issue[1].

Some of the firmware blobs are provided under the 'Synapse' license
including a clause that can (at least theoretically) be used as a
'killswitch'. To make sure users are aware of this, we guard this
behind a license flag.

Users of this recipe will need to add to their configuration:

    LICENSE_FLAGS_ACCEPTED = "synaptics-killswitch"

[1] https://github.com/RPi-Distro/firmware-nonfree/issues/29

Signed-off-by: Andrei Gherzan <andrei.gherzan@huawei.com>
(cherry picked from commit 7c31b9b4b595c3e0f932391e07a5258a9f966e4c)